requirements.txt
django~=5.1.0
psycopg[binary]~=3.2.0
django-filter~=24.3.0
djangorestframework-gis~=1.0.0
djangorestframework~=3.15.0
rasterio~=1.4.3
pyproj~=3.7.0
Pillow~=11.1.0
